move components to a proper folder
