+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_monitoring_ui/components/health_overview/index.ts
